Rename routers
