+++ routes/inventoryRouter.js
